# docs/swagger.yaml
basePath: /
definitions:
...
  CurrencyControl_internal_service_ports.LoginResult:
    properties:
      access_token:
        type: string
      access_token_expires_at:
        type: string
      login:
        type: string
      message:
        type: string
      refresh_token:
        type: string
      refresh_token_expires_at:
        type: string
      session:
        $ref: '#/definitions/domain.Session'
      status:
        $ref: '#/definitions/CurrencyControl_internal_service_ports.LoginStatus'
      token:
        type: string
    type: object
  CurrencyControl_internal_service_ports.LoginStatus:
    enum:
    - active
    - no_role
    type: string
    x-enum-varnames:
    - StatusActive
    - StatusNoRole
  delivery_http.AccessRequestResponse:
    properties:
      branch_id:
        type: integer
      branch_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      login:
        type: string
      message:
        type: string
      reviewed_at:
        type: string
      role:
        type: string
      session_token:
        type: string
      status:
        type: string
    type: object
...
  delivery_http.CommonError:
    properties:
      error:
        type: string
    type: object
...
  delivery_http.loginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  delivery_http.refreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  delivery_http.requestAccessBody:
    properties:
      branch_id:
        description: ID филиала (из списка GET /auth/branches)
        example: 5100
        type: integer
      role:
        enum:
        - operator
        - branch_head
        - currency_control
        - compliance
        - internal_audit
        example: operator
        type: string
    type: object
  domain.AccessRequest:
    properties:
      branch_id:
        type: integer
      branch_name:
        type: string
      created_at:
        type: string
      id:
        type: integer
      login:
        type: string
      reviewed_at:
        type: string
      role:
        type: string
      session_token:
        type: string
      status:
        type: string
    type: object
...
  domain.Session:
    properties:
      branch_id:
        type: integer
      created_at:
        type: string
      expires_at:
        type: string
      id:
        type: integer
      login:
        type: string
      role:
        type: string
      token:
        type: string
    type: object
...
host: localhost:8088
info:
  contact: {}
  description: API сервера для системы валютного контроля.
  title: Currency Control API
  version: "1.0"
paths:
  /api/access-requests:
    get:
      description: 'Возвращает заявки со статусом pending. Доступно: Комплаенс, Администратор.'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AccessRequest'
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Список запросов на доступ (ожидающих)
      tags:
      - AccessRequests
  /api/access-requests/{request_id}/approve:
    post:
      description: 'Одобряет заявку и создаёт JWT access/refresh токены. Доступно:
        Комплаенс, Администратор.'
      parameters:
      - description: ID заявки
        in: path
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CurrencyControl_internal_service_ports.LoginResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Одобрить запрос на доступ
      tags:
      - AccessRequests
  /api/access-requests/{request_id}/reject:
    post:
      description: 'Отклоняет заявку на доступ. Доступно: Комплаенс, Администратор.'
      parameters:
      - description: ID заявки
        in: path
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Отклонить запрос на доступ
      tags:
      - AccessRequests
...
  /auth/access-requests/{request_id}/status:
    get:
      description: Пользователь опрашивает статус своей заявки. Если одобрено — возвращает
        токен сессии.
      parameters:
      - description: ID заявки (из POST /auth/request-access)
        in: path
        name: request_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Проверить статус запроса на доступ
      tags:
      - Auth
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет логин/пароль через AD.
      parameters:
      - description: Логин и пароль
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/delivery_http.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Service Unavailable
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход в систему
      tags:
      - Auth
  /auth/logout:
    post:
      description: Удаляет текущую refresh-сессию пользователя.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выход из системы
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Принимает refresh token и возвращает новый access token.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/delivery_http.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CurrencyControl_internal_service_ports.LoginResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      summary: Обновить access token
      tags:
      - Auth
  /auth/request-access:
    post:
      consumes:
      - application/json
      description: Новый пользователь, прошедший AD-проверку, выбирает свой филиал
        и роль из существующих.
      parameters:
      - description: Выбор филиала и роли (выберите из существующих вариантов)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/delivery_http.requestAccessBody'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/delivery_http.AccessRequestResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/delivery_http.CommonError'
      security:
      - ApiKeyAuth: []
      summary: Запрос на доступ (для новых пользователей)
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    description: Введите Bearer access_token (полученный при /auth/login или /auth/refresh)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"